# api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Hackathon 2026 智能评审Agent系统 API
  description: |
    为AI编程大赛提供自动化的作品提交、评审任务调度、评分结果查询和排行榜服务。
    所有时间戳字段使用ISO 8601格式。
  version: 1.0.0
  contact:
    name: Hackathon Tech Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local Development
  - url: https://hackathon2026.example.com/api/v1
    description: Production
paths:
...
  /teams:
    post:
      summary: 注册参赛队伍
      operationId: createTeam
      tags: [Teams]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreate'
      responses:
        '201':
          description: 队伍创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          description: 参数校验失败

    get:
      summary: 获取队伍列表
      operationId: listTeams
      tags: [Teams]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: 队伍列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTeams'
  /teams/{team_id}:
    get:
      summary: 获取队伍详情
      operationId: getTeam
      tags: [Teams]
      parameters:
        - name: team_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 队伍详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamWithSubmissions'

  # ============================================================
  # 赛题管理
  # ============================================================
...
  /submissions:
    post:
      summary: 提交参赛作品（Multipart文件上传）
      operationId: createSubmission
      tags: [Submissions]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [team_id, question_id, file]
              properties:
                team_id:
                  type: string
                  format: uuid
                  description: 队伍ID
                question_id:
                  type: integer
                  description: 赛题编号(1-7)
                file:
                  type: string
                  format: binary
                  description: 作品ZIP/TAR.GZ文件，最大500MB
                description:
                  type: string
                  description: 作品说明（可选）
      responses:
        '201':
          description: 提交成功，评审任务已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
        '400':
          description: 文件格式错误或参数无效
        '413':
          description: 文件过大

    get:
      summary: 查询提交列表（支持过滤）
      operationId: listSubmissions
      tags: [Submissions]
      parameters:
        - name: team_id
          in: query
          schema: { type: string, format: uuid }
        - name: question_id
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed]
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: 提交列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSubmissions'
  /submissions/{submission_id}:
    get:
      summary: 获取作品详情及评审状态
      operationId: getSubmission
      tags: [Submissions]
      parameters:
        - name: submission_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 作品详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionDetail'
  /submissions/{submission_id}/retry:
    post:
      summary: 重新触发评审（用于失败或手动重试）
      operationId: retrySubmissionReview
      tags: [Submissions]
      parameters:
        - name: submission_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '202':
          description: 评审任务已重新提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTask'

  # ============================================================
  # 评审结果查询
  # ============================================================
...
  /tasks/{task_id}:
    get:
      summary: 查询评审任务状态
      operationId: getTaskStatus
      tags: [System]
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewTask'
components:
  schemas:
...
    TeamCreate:
      type: object
      required: [team_code, team_name]
      properties:
        team_code:
          type: string
          example: "T001"
        team_name:
          type: string
          example: "Code Masters"
        members:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              role: { type: string }
              email: { type: string, format: email }
        contact_email:
          type: string
          format: email
    Team:
      type: object
      properties:
        id: { type: string, format: uuid }
        team_code: { type: string }
        team_name: { type: string }
        members: { type: array }
        contact_email: { type: string }
        created_at: { type: string, format: date-time }
    TeamWithSubmissions:
      allOf:
        - $ref: '#/components/schemas/Team'
        - type: object
          properties:
            submissions:
              type: array
              items:
                $ref: '#/components/schemas/SubmissionResponse'
    PaginatedTeams:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        items:
          type: array
          items:
            $ref: '#/components/schemas/Team'
...
    SubmissionResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        team_id: { type: string, format: uuid }
        question_id: { type: integer }
        submission_version: { type: integer }
        file_path: { type: string }
        file_size: { type: integer }
        status:
          type: string
          enum: [pending, running, completed, failed]
        submitted_at: { type: string, format: date-time }
        review_task_id: { type: string, format: uuid, nullable: true }
        estimated_completion: { type: string, format: date-time, nullable: true }
    PaginatedSubmissions:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
        items:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionResponse'
    SubmissionDetail:
      allOf:
        - $ref: '#/components/schemas/SubmissionResponse'
        - type: object
          properties:
            metadata:
              type: object
              description: "解析出的项目元数据"
            completed_at: { type: string, format: date-time, nullable: true }
            final_score:
              type: object
              nullable: true
              properties:
                total_score: { type: number }
                confidence: { type: number }
                human_review_required: { type: boolean }
                human_reviewed: { type: boolean }
            tasks:
              type: array
              items:
                $ref: '#/components/schemas/ReviewTask'
    ReviewTask:
      type: object
      properties:
        id: { type: string, format: uuid }
        submission_id: { type: string, format: uuid }
        task_type:
          type: string
          enum: [full_review, sandbox_test, llm_review]
        status:
          type: string
          enum: [pending, running, success, failure, retry]
        started_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
        error_message: { type: string, nullable: true }
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []